WIP: Smaract MCS2 via Python-API
WIP: Adding Smaract XYZ-Stage with MCS2 controller, using Smaracts new Python-API. X and Y seem to work already, but the Z axis freezes after one move.
It is not clear why this happens, as on the hardware-side the axes are all identical. There may be some difference in the way cockpit handles the Z-axis.
Alternatively,
@cockpit.util.threads.callInNewThread
def sendXYZPositionUpdates(self):
might be at fault.
It's current design is really simple, as the more "elegant" solutions created problems even with X and Y.